backend_python/requirements.txt
Flask
flask-cors
torch
Pillow
requests
python-dotenv
transformers
ultralytics
numpy
